some refactoring of niuzao module
